// src/components/console/UploadFile/enum.ts
export enum NoteType {
  NOTA1 = "N1",
  NOTA2 = "N2",
  NOTA3 = "N3"
}

export enum RequiredAttributes {
  DISCIPLINA = "DISCIPLINA",
  PROFESSORES = "PROFESSORES",
  NOME = "NOME",
  STATUS = "STATUS",
  NOTA1 = "NOTA1",
  NOTA2 = "NOTA2",
  NOTA3 = "NOTA3",
  NOTASUBS = "NOTASUBS",
  MEDIAFINAL = "MEDIAFINAL",
  FALTAS_TOTAL = "FALTAS_TOTAL",
  MATRICULA = "MATRICULA",
  TYPE_FOR_NOTE = "TIPO_PARA_NOTA",
  TYPE_FOR_FOULS = "TIPO_PARA_FALTAS"
}

export enum Subject {
  NAME = "name",
  HOURS = "hours"
}

export enum ToastType {
  SUCCESS = "success",
  WARN = "warning",
  ERROR = "error"
}